Renamed user_models file ==> models + Added database check model in this new models file + Applied migrations
